refactor(cli): adopt package manager command namespace
