+++ packages/cli-defaults/package.json
@@ -40,9 +40,9 @@
     "@backstage/cli-module-maintenance": "workspace:^",
     "@backstage/cli-module-migrate": "workspace:^",
     "@backstage/cli-module-new": "workspace:^",
+    "@backstage/cli-module-package-manager-yarn": "workspace:^",
     "@backstage/cli-module-test-jest": "workspace:^",
-    "@backstage/cli-module-translations": "workspace:^",
-    "@backstage/cli-module-yarn": "workspace:^"
+    "@backstage/cli-module-translations": "workspace:^"
   },
   "devDependencies": {
     "@backstage/cli": "workspace:^"
