+++ packages/cli-defaults/src/index.ts
@@ -23,9 +23,9 @@ import lint from '@backstage/cli-module-lint';
 import maintenance from '@backstage/cli-module-maintenance';
 import migrate from '@backstage/cli-module-migrate';
 import newModule from '@backstage/cli-module-new';
+import packageManagerYarn from '@backstage/cli-module-package-manager-yarn';
 import testJest from '@backstage/cli-module-test-jest';
 import translations from '@backstage/cli-module-translations';
-import yarn from '@backstage/cli-module-yarn';
 
 /**
  * The default set of CLI modules for the Backstage CLI.
@@ -43,7 +43,7 @@ export default [
   maintenance,
   migrate,
   newModule,
+  packageManagerYarn,
   testJest,
   translations,
-  yarn,
 ];
