+++ packages/cli-module-package-manager-yarn/cli-report.md
@@ -1,0 +1,40 @@
+## CLI Report file for "@backstage/cli-module-package-manager-yarn"
+
+> Do not edit this file. It is a report generated by `yarn build:api-reports`
+
+### `backstage-cli-module-package-manager-yarn`
+
+```
+Usage: @backstage/cli-module-package-manager-yarn [flags...]
+
+Options:
+  -V, --version
+  -h, --help
+
+Commands:
+  help
+  pm
+```
+
+### `backstage-cli-module-package-manager-yarn pm`
+
+```
+Usage: @backstage/cli-module-package-manager-yarn pm [flags...]
+
+Options:
+  -V, --version
+  -h, --help
+
+Commands:
+  help
+  verify-patches
+```
+
+### `backstage-cli-module-package-manager-yarn pm verify-patches`
+
+```
+Usage: @backstage/cli-module-package-manager-yarn pm verify-patches [flags...]
+
+Options:
+  -h, --help
+```
