+++ packages/cli-module-package-manager-yarn/report.api.md
@@ -1,4 +1,4 @@
-## API Report File for "@backstage/cli-module-yarn"
+## API Report File for "@backstage/cli-module-package-manager-yarn"
 
 > Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
