--- packages/cli-module-yarn/catalog-info.yaml
@@ -1,10 +1,0 @@
-apiVersion: backstage.io/v1alpha1
-kind: Component
-metadata:
-  name: backstage-cli-module-yarn
-  title: '@backstage/cli-module-yarn'
-  description: CLI module for Backstage CLI
-spec:
-  lifecycle: experimental
-  type: backstage-cli-module
-  owner: tooling-maintainers
