--- packages/cli-module-yarn/cli-report.md
@@ -1,40 +1,0 @@
-## CLI Report file for "@backstage/cli-module-yarn"
-
-> Do not edit this file. It is a report generated by `yarn build:api-reports`
-
-### `backstage-cli-module-yarn`
-
-```
-Usage: @backstage/cli-module-yarn [flags...]
-
-Options:
-  -V, --version
-  -h, --help
-
-Commands:
-  help
-  repo
-```
-
-### `backstage-cli-module-yarn repo`
-
-```
-Usage: @backstage/cli-module-yarn repo [flags...]
-
-Options:
-  -V, --version
-  -h, --help
-
-Commands:
-  help
-  verify-yarn-patches
-```
-
-### `backstage-cli-module-yarn repo verify-yarn-patches`
-
-```
-Usage: @backstage/cli-module-yarn repo verify-yarn-patches [flags...]
-
-Options:
-  -h, --help
-```
